Create Password_Checker
Pwned Passwords is a free online repo which stores passwords leaked in data breaches. This script safely checks your passwords on your local machine by making a call to the Pwned Passwords API. This is achieved by sending an encrypted section of the password to the API, the encrypted is done via the SHA1 cryptographic hash function. The API returns a list of hacked passwords which match the section sent to the API.  Your full password is checked against this list locally so your full password information never leaves your machine.